split execution component out of agent
