+++ src/radical/pilot/agent/__init__.py
@@ -6,4 +6,5 @@ __license__   = "MIT"
 from .rm        import RM
 from .lm        import LaunchMethod as LM
 from .scheduler import Scheduler
+from .executing import Executing
